requirements.txt
spacy==3.6.0
pdfminer.six==20221105
requests==2.31.0
